requirements.txt
telethon 
fastapi
uvicorn 
python-dotenv
